Create ConstraintSolver interface
